Remove :: bindings and other small TS tweaks (#711)

+++ packages/cx/src/data/View.js
@@ -1,48 +1,40 @@
-import {Binding} from './Binding';
-import {isArray} from '../util/isArray';
-import {isDefined} from "../util/isDefined";
-import {StoreRef} from "./StoreRef";
+import { Binding } from "./Binding";
+import { isArray } from "../util/isArray";
+import { isDefined } from "../util/isDefined";
+import { StoreRef } from "./StoreRef";
 
 export class View {
-
    constructor(config) {
       Object.assign(this, config);
       this.cache = {
-         version: -1
+         version: -1,
       };
-      if (this.store)
-         this.setStore(this.store);
+      if (this.store) this.setStore(this.store);
    }
 
    getData() {
-      throw new Error('abstract method');
+      throw new Error("abstract method");
    }
 
    init(path, value) {
-      if (path instanceof Binding)
-         path = path.path;
-      else if (typeof path == 'object' && path != null) {
+      if (path instanceof Binding) path = path.path;
+      else if (typeof path == "object" && path != null) {
          var changed = false;
          for (var key in path)
-            if (path.hasOwnProperty(key) && this.get(key) === undefined && this.setItem(key, path[key]))
-               changed = true;
+            if (path.hasOwnProperty(key) && this.get(key) === undefined && this.setItem(key, path[key])) changed = true;
          return changed;
       }
 
-      if (this.get(path) === undefined)
-         return this.setItem(path, value);
+      if (this.get(path) === undefined) return this.setItem(path, value);
 
       return false;
    }
 
    set(path, value) {
-      if (path instanceof Binding)
-         path = path.path;
-      else if (typeof path == 'object' && path != null) {
+      if (path instanceof Binding) path = path.path;
+      else if (typeof path == "object" && path != null) {
          var changed = false;
-         for (var key in path)
-            if (path.hasOwnProperty(key) && this.setItem(key, path[key]))
-               changed = true;
+         for (var key in path) if (path.hasOwnProperty(key) && this.setItem(key, path[key])) changed = true;
          return changed;
       }
       return this.setItem(path, value);
@@ -62,53 +54,45 @@ export class View {
 
    //protected
    setItem(path, value) {
-      if (this.store)
-         return this.store.setItem(path, value);
+      if (this.store) return this.store.setItem(path, value);
 
-      throw new Error('abstract method');
+      throw new Error("abstract method");
    }
 
    delete(path) {
-      if (path instanceof Binding)
-         path = path.path;
-      else if (arguments.length > 1)
-         path = Array.from(arguments);
+      if (path instanceof Binding) path = path.path;
+      else if (arguments.length > 1) path = Array.from(arguments);
 
-      if (isArray(path))
-         return path.map(arg => this.deleteItem(arg)).some(Boolean);
+      if (isArray(path)) return path.map((arg) => this.deleteItem(arg)).some(Boolean);
 
       return this.deleteItem(path);
    }
 
    //protected
    deleteItem(path) {
-      if (this.store)
-         return this.store.deleteItem(path);
+      if (this.store) return this.store.deleteItem(path);
 
-      throw new Error('abstract method');
+      throw new Error("abstract method");
    }
 
    clear() {
-      if (this.store)
-         return this.store.clear();
+      if (this.store) return this.store.clear();
 
-      throw new Error('abstract method');
+      throw new Error("abstract method");
    }
 
    get(path) {
       let storeData = this.getData();
 
-      if (arguments.length > 1)
-         path = Array.from(arguments);
+      if (arguments.length > 1) path = Array.from(arguments);
 
-      if (isArray(path))
-         return path.map(arg => Binding.get(arg).value(storeData));
+      if (isArray(path)) return path.map((arg) => Binding.get(arg).value(storeData));
 
       return Binding.get(path).value(storeData);
    }
 
    toggle(path) {
-      return this.set(path, !this.get(path))
+      return this.set(path, !this.get(path));
    }
 
    update(path, updateFn, ...args) {
@@ -117,49 +101,41 @@ export class View {
 
    batch(callback) {
       let dirty = this.silently(callback);
-      if (dirty)
-         this.notify();
+      if (dirty) this.notify();
       return dirty;
    }
 
    silently(callback) {
-      if (this.store)
-         return this.store.silently(callback);
+      if (this.store) return this.store.silently(callback);
 
-      throw new Error('abstract method');
+      throw new Error("abstract method");
    }
 
    notify(path) {
-      if (this.notificationsSuspended)
-         this.dirty = true;
-      else
-         this.doNotify(path);
+      if (this.notificationsSuspended) this.dirty = true;
+      else this.doNotify(path);
    }
 
    doNotify(path) {
-      if (this.store)
-         return this.store.notify(path);
+      if (this.store) return this.store.notify(path);
 
-      throw new Error('abstract method');
+      throw new Error("abstract method");
    }
 
    subscribe(callback) {
-      if (this.store)
-         return this.store.subscribe(callback);
+      if (this.store) return this.store.subscribe(callback);
 
-      throw new Error('abstract method');
+      throw new Error("abstract method");
    }
 
    load(data) {
-      return this.batch(store => {
-         for (var key in data)
-            store.set(key, data[key]);
+      return this.batch((store) => {
+         for (var key in data) store.set(key, data[key]);
       });
    }
 
    dispatch(action) {
-      if (this.store)
-         return this.store.dispatch(action);
+      if (this.store) return this.store.dispatch(action);
 
       throw new Error("The underlying store doesn't support dispatch.");
    }
@@ -174,25 +150,24 @@ export class View {
    }
 
    ref(path, defaultValue) {
-      if (isDefined(defaultValue))
-         this.init(path, defaultValue);
+      if (isDefined(defaultValue)) this.init(path, defaultValue);
       return StoreRef.create({
          store: this,
-         path
+         path,
       });
    }
 
    getMethods() {
       return {
-         getData: ::this.getData,
-         set: ::this.set,
-         get: ::this.get,
-         update: ::this.update,
-         delete: ::this.delete,
-         toggle: ::this.toggle,
-         init: ::this.init,
-         ref: ::this.ref
-      }
+         getData: this.getData.bind(this),
+         set: this.set.bind(this),
+         get: this.get.bind(this),
+         update: this.update.bind(this),
+         delete: this.delete.bind(this),
+         toggle: this.toggle.bind(this),
+         init: this.init.bind(this),
+         ref: this.ref.bind(this),
+      };
    }
 }
